+++ metrics/domain/src/main/kotlin/me/proton/core/metrics/domain/repository/MetricsRepository.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) 2020 Proton Technologies AG
+ * This file is part of Proton Technologies AG and ProtonCore.
+ *
+ * ProtonCore is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * ProtonCore is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with ProtonCore.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package me.proton.core.metrics.domain.repository
+
+import me.proton.core.domain.entity.UserId
+import me.proton.core.metrics.domain.entity.Metrics
+
+public interface MetricsRepository {
+
+    /**
+     * Post metrics to the backend.
+     * If the [userId] is provided, it will be used
+     * to authenticate the API call.
+     */
+    public suspend fun post(userId: UserId?, metrics: Metrics)
+
+}