+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### New
+- Metrics modules (Dagger, Data, Domain).
+
 ## [7.1.0]
 
 ### New